requirements.txt
numpy==1.24.3
Flask==2.2.2
Werkzeug==2.2.2
redis==3.5.3
openai==1.12.0
PyGithub==2.2.0
markdown-it-py==3.0.0
python-dotenv==0.19.1
urllib3==1.26.7